Add `unwindIgnoreEmptyArrays` option.
In order to convert the generated CSV back to the most accurate JSON
possible, the module must still insert `[]` for unwound empty array
values. However, there are some legitimate cases where users want to be
able to have the module ignore these values so that the generated CSV is
a bit more user-friendly, depending on the data being converted. This
new option allows users the ability to achieve this without impacting
existing use cases, as the user has opted to specify an option which may
impact the ability to convert the CSV back to JSON later on.

Fixes #168